chore: refactoring, renaming, testing
